pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "llm-pipeline"
version = "0.1.0"
description = "Declarative LLM pipeline orchestration framework"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "pydantic>=2.0",
    "sqlmodel>=0.0.14",
    "sqlalchemy>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
gemini = ["google-generativeai>=0.3.0"]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
